Extract `SessionController`
...for use in all session-related activities
