+++ PocketKit/Sources/PocketKit/Environment.swift
@@ -6,14 +6,6 @@ import SwiftUI
 import Sync
 
 
-private struct AccessTokenStoreKey: EnvironmentKey {
-    static var defaultValue: AccessTokenStore = Services.shared.accessTokenStore
-}
-
-private struct AuthorizationClientKey: EnvironmentKey {
-    static var defaultValue: AuthorizationClient = Services.shared.authClient
-}
-
 private struct SourceKey: EnvironmentKey {
     static var defaultValue = Services.shared.source
 }
@@ -23,16 +15,6 @@ private struct CharacterDirectionKey: EnvironmentKey {
 }
 
 extension EnvironmentValues {
-    var accessTokenStore: AccessTokenStore {
-        get { self[AccessTokenStoreKey.self] }
-        set { self[AccessTokenStoreKey.self] = newValue }
-    }
-
-    var authorizationClient: AuthorizationClient {
-        get { self[AuthorizationClientKey.self] }
-        set { self[AuthorizationClientKey.self] = newValue }
-    }
-
     var source: Source {
         get { self[SourceKey.self] }
         set { self[SourceKey.self] = newValue }
